--- src/components/header/Header.module.css
@@ -1,136 +1,0 @@
-.headerContainer {
-  background-color: white;
-  border-bottom: 1px rgba(0, 0, 0, 0.1) solid;
-  width: 100%;
-  height: 70px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: white;
-  z-index: 10;
-}
-.headerWrapper {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: fit-content;
-  column-gap: 3em;
-  display: flex;
-}
-.logoSVGwrapper {
-  z-index: 0;
-  position: absolute;
-  left: 22px;
-  background-color: var(--finedblue);
-  width: 40px;
-  height: 40px;
-  border-radius: 5px;
-}
-.logoWrapper {
-  position: relative;
-  width: fit-content;
-  padding: 10px 25px 10px 25px;
-  background-color: white;
-  display: flex;
-  column-gap: 14px;
-  justify-content: center;
-  align-items: center;
-  .logoText {
-    transition: 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-    color: var(--finedblue);
-    font-size: 1.8rem;
-    font-weight: 570;
-    z-index: 1;
-    margin: 0;
-  }
-  .logoSVGwrapper {
-    transition: 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-  }
-  .logosvg {
-    z-index: 10;
-  }
-  &:hover {
-    cursor: pointer;
-    color: white;
-    .logoText {
-      transition: 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-      color: white;
-    }
-    .logoSVGwrapper {
-      transition: 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-      width: 100%;
-      height: 100%;
-      left: 0;
-    }
-  }
-}
-.list {
-  display: flex;
-  justify-content: end;
-  flex-direction: row;
-  column-gap: 2em;
-  width: 10em;
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-.link {
-  transition: opacity 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-  color: var(--black1);
-  opacity: 0.6;
-  &:hover {
-    transition: opacity 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-    opacity: 1;
-  }
-}
-.search {
-  width: 10em;
-}
-
-.hamburgerPart {
-  width: 16px;
-  height: 3px;
-  background-color: #9e9e9e;
-  border-radius: 2px;
-  transition: color 0.3s, cubic-bezier(0.215, 0.61, 0.355, 1);
-}
-.sideBarMinimizer {
-  transition: 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-  display: flex;
-  flex-direction: column;
-  row-gap: 2px;
-  justify-content: center;
-  align-items: center;
-  position: absolute;
-  left: 2em;
-  background-color: #f5f5f5;
-  width: 55px;
-  height: 32px;
-  border: 1px solid var(--inputborder);
-  border-radius: 30px;
-}
-
-.sideBarMinimizer:hover {
-  transition: 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-  cursor: pointer;
-  background-color: #f0f0f0;
-  .hamburgerPart {
-    transition: 0.3s cubic-bezier(0.215, 0.61, 0.355, 1);
-    background-color: var(--black3);
-  }
-}
-
-@media only screen and (max-width: 1280px) {
-  .headerWrapper {
-    width: 95%;
-    padding-left: 5em;
-    box-sizing: border-box;
-    justify-content: space-between;
-  }
-  .list {
-    display: none;
-  }
-  .search {
-    width: 30em;
-  }
-}
